DELIA-49068 - dtcpmgr changes
Reason for change: dtcpmgr changes performed
Test Procedure: Sanity Testing
Risks: Low

(cherry picked from commit 65bace949ed25ac96fe4c0f38de42df2a31b4b82)

+++ include/dtcpmgr.h
@@ -167,7 +167,7 @@ typedef unsigned char  BOOLEAN;
  *
  * An opaque handle to a DTCP-IP source/sink session.
  */
-typedef unsigned long long DTCP_SESSION_HANDLE;
+typedef unsigned long DTCP_SESSION_HANDLE;
 
 /**
  * @brief DTCP-IP device types.
